+++ packages/tools/examples/dynamicallyAddAnnotations/toolSpecificUI.ts
@@ -1,26 +1,11 @@
+import { createAngleToolUI } from './angleToolUI';
+import { createArrowAnnotateToolUI } from './arrowAnnotateToolUI';
+import { createEllipseROIToolUI } from './ellipticalROIToolUI';
 import { createLengthToolUI } from './lengthToolUI';
 import { createProbeToolUI } from './probeToolUI';
 import { createRectangleROIToolUI } from './rectangleROIToolUI';
 import { createCircleROIToolUI } from './circleROIToolUI';
 import { createSplineROIToolUI } from './splineROIToolUI';
-
-export const STACK_VIEWPORT_ID = 'viewport-stack';
-export const VOLUME_VIEWPORT_ID = 'viewport-volume';
-
-export const typeToIdMap = {
-  stack: STACK_VIEWPORT_ID,
-  volume: VOLUME_VIEWPORT_ID,
-} as const;
-
-export const typeToStartIdMap = {
-  canvas: 'canvas-start',
-  image: 'image-start',
-} as const;
-
-export const typeToEndIdMap = {
-  canvas: 'canvas-end',
-  image: 'image-end',
-} as const;
 
 interface ToolUIConfig {
   toolName: string;
@@ -37,6 +22,15 @@ function createToolUI(toolName: string, config: ToolUIConfig): ToolUI | null {
   let forms: HTMLElement[] = [];
 
   switch (toolName) {
+    case 'Angle':
+      forms = [createAngleToolUI()];
+      break;
+    case 'ArrowAnnotate':
+      forms = [createArrowAnnotateToolUI()];
+      break;
+    case 'EllipticalROI':
+      forms = [createEllipseROIToolUI()];
+      break;
     case 'Length':
       forms = [createLengthToolUI()];
       break;