+++ utils/ExampleRunner/rules-examples.js
@@ -2,17 +2,52 @@ const autoprefixer = require('autoprefixer');
 const excludeNodeModulesExcept = require('../../.webpack/excludeNodeModulesExcept.js');
 
 const exclude = excludeNodeModulesExcept([]);
+const prod = process.env.NODE_ENV === 'production';
 
 module.exports = [
   {
-    test: /\.(mjs|ts|tsx|js|jsx)?$/,
+    test: /\.(j|t)s$/,
     exclude,
-    loader: 'babel-loader',
+    loader: 'builtin:swc-loader',
     options: {
-      // Find babel.config.js in monorepo root
-      // https://babeljs.io/docs/en/options#rootmode
-      rootMode: 'upward',
-      envName: 'development',
+      jsc: {
+        parser: {
+          syntax: 'typescript',
+        },
+        transform: {
+          react: {
+            runtime: 'automatic',
+            development: !prod,
+            refresh: !prod,
+          },
+        },
+      },
+      env: {
+        targets: 'Chrome >= 48',
+      },
+    },
+  },
+  {
+    test: /\.(j|t)sx$/,
+    exclude,
+    loader: 'builtin:swc-loader',
+    options: {
+      jsc: {
+        parser: {
+          syntax: 'typescript',
+          tsx: true,
+        },
+        transform: {
+          react: {
+            runtime: 'automatic',
+            development: !prod,
+            refresh: !prod,
+          },
+        },
+      },
+      env: {
+        targets: 'Chrome >= 48',
+      },
     },
   },
   {
@@ -32,12 +67,8 @@ module.exports = [
     ],
   },
   {
-    test: /\.png$/i,
-    use: [
-      {
-        loader: 'url-loader',
-      },
-    ],
+    test: /\.(png|jpe?g|gif)$/i,
+    type: 'asset/resource',
   },
   {
     test: /\.wasm$/,